Updated with Compose

+++ app/src/main/java/com/example/rangtech/models/AnswerChoice.kt
@@ -3,5 +3,5 @@ package com.example.rangtech.models
 data class AnswerChoice(
     val id: Int,
     val name: String,
-    val score: Float
+    val score: Double
 )
